# openapi.yaml
openapi: 3.1.0
info:
  version: 0.2.4
  title: Owncast APIs
  description: |-
    Internal
    The APIs required to make the Owncast web interface and internal functionality work. These APIs will never reach stability and will always change as Owncast features and functionality change. This means these endpoints are not good options to build your own tooling and features on top of. You can, however, of course can if that risk is acceptable to you.

    Admin
    The APIs used to configure the Owncast server, as well as power the web admin interface. These APIs are protected by the admin password, using HTTP Basic authentication. They should not be used by external tools or integrations as it would require you to give your admin password to a third party.
  contact:
    name: Gabe Kangas
    email: gabek@real-ity.com
    url: 'http://owncast.online'
  license:
    url: 'https://github.com/owncast/owncast/blob/develop/LICENSE'
    name: MIT
  x-logo:
    url: >-
...
servers:
  - url: 'http://localhost:8080/api'
    description: Localhost port 8080
paths:
...
  /yp:
    get:
      summary: Get the YP protocol data
      operationId: GetYPResponse
      tags: ['Internal']
      responses:
        '200':
          description: The current YP protocol configuration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/YPDetails'
        '404':
          $ref: '#/components/responses/404'
...
  /admin/status:
    get:
      summary: Get current inboard broadcaster
      operationId: StatusAdmin
      tags: ['Internal', 'Admin']
      security:
        - BasicAuth: []
      responses:
        '200':
          description: Gives the status of the system
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminStatus'
        '400':
          $ref: '#/components/responses/400'
        '401':
          $ref: '#/components/responses/401BasicAuth'
        default:
          $ref: '#/components/responses/Default'
    options:
      operationId: StatusAdminOptions
      x-internal: true
      tags: ['Objects']
      responses:
        '204':
          $ref: '#/components/responses/204'
...
  /integrations/status:
    get:
      summary: Get the server's status
      operationId: ExternalGetStatus
      tags: ['External']
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        '401':
          $ref: '#/components/responses/401BasicAuth'
        '404':
          $ref: '#/components/responses/404'
        default:
          $ref: '#/components/responses/Default'
...
components:
  schemas:
...
    BaseAPIResponse:
      type: object
      description: Simple API response
      properties:
        message:
          type: string
        success:
          type: boolean
    Status:
      type: object
      description: Response for status
      properties:
        serverTime:
          type: string
        lastConnectTime:
          type: ['string', 'null']
        lastDisconnectTime:
          type: ['string', 'null']
        versionNumber:
          type: string
        viewerCount:
          type: integer
        online:
          type: boolean
        streamTitle:
          type: string
...
    SocialHandle:
      type: object
      properties:
        platform:
          type: string
        url:
          type: string
        icon:
          type: string
...
    YPDetails:
      type: object
      properties:
        lastConnectTime:
          type: string
        name:
          type: string
        description:
          type: string
        streamTitle:
          type: string
        logo:
          type: string
        tags:
          type: array
          items:
            type: string
        social:
          type: array
          items:
            $ref: '#/components/schemas/SocialHandle'
        viewerCount:
          type: integer
        overallMaxViewerCount:
          type: integer
        sessionMaxViewerCount:
          type: integer
        nsfw:
          type: boolean
        online:
          type: boolean
...
    Broadcaster:
      type: object
      properties:
        time:
          type: string
          format: date-time
        remoteAddr:
          type: string
        streamDetails:
          $ref: '#/components/schemas/InboundStreamDetails'
    InboundStreamDetails:
      type: object
      properties:
        videoCodec:
          type: string
        audioCodec:
          type: string
        encoder:
          type: string
        width:
          type: integer
        height:
          type: integer
        videoBitrate:
          type: integer
        audioBitrate:
          type: integer
        framerate:
          type: number
          format: float
    CurrentBroadcast:
      type: object
      properties:
        outputSettings:
          type: array
          items:
            $ref: '#/components/schemas/StreamOutputVariant'
        latencyLevel:
          $ref: '#/components/schemas/LatencyLevel'
    StreamOutputVariant:
      type: object
      properties:
        name:
          type: string
        videoPassthrough:
          type: boolean
        audioPassthrough:
          type: boolean
        videoBitrate:
          type: integer
        audioBitrate:
          type: integer
        scaledWidth:
          type: integer
        scaledHeight:
          type: integer
        framerate:
          type: integer
        cpuUsageLevel:
          type: integer
    LatencyLevel:
      type: object
      properties:
        level:
          type: integer
    StreamHealthOverview:
      type: object
      properties:
        message:
          type: string
        healthPercentage:
          type: integer
        representation:
          type: integer
        healthy:
          type: boolean
...
    AdminStatus:
      type: object
      properties:
        broadcaster:
          $ref: '#/components/schemas/Broadcaster'
        currentBroadcast:
          $ref: '#/components/schemas/CurrentBroadcast'
        health:
          anyOf:
            - $ref: '#/components/schemas/StreamHealthOverview'
            - type: 'null'
        streamTitle:
          type: string
        versionNumber:
          type: string
        viewerCount:
          type: integer
        overallPeakViewerCount:
          type: integer
        sessionPeakViewerCount:
          type: integer
        online:
          type: boolean
...
  headers:
    'Access-Control-Allow-Origin':
      schema:
        type: string
    'Access-Control-Allow-Credentials':
      schema:
        type: string
        format: boolean
    'Access-Control-Allow-Headers':
      schema:
        type: string
    'WWW-Authenticate':
      schema:
        type: string
    'X-Content-Type-Options':
      schema:
        type: string
...
  responses:
    Default:
      description: Unexpected error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/BaseAPIResponse'
    '204':
      description: No Content
...
    '400':
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/BaseAPIResponse'
...
    '401BasicAuth':
      description: Basic auth failure
      headers:
        'Access-Control-Allow-Origin':
          $ref: '#/components/headers/Access-Control-Allow-Origin'
        'Access-Control-Allow-Credentials':
          $ref: '#/components/headers/Access-Control-Allow-Credentials'
        'Access-Control-Allow-Headers':
          $ref: '#/components/headers/Access-Control-Allow-Headers'
        'WWW-Authenticate':
          $ref: '#/components/headers/WWW-Authenticate'
        'X-Content-Type-Options':
          $ref: '#/components/headers/X-Content-Type-Options'
      content:
        text/plain:
          schema:
            type: string
...
    '404':
      description: Not found
...
  securitySchemes:
    BasicAuth:
      type: http
      scheme: basic
    BearerAuth:
      type: http
      scheme: bearer
...